chore: rm unused boltManager file
