+++ bolt-cli/src/commands/operators/eigenlayer.rs
@@ -15,13 +15,15 @@ use tracing::{info, warn};
 use crate::{
     cli::{Chain, EigenLayerSubcommand},
     common::{
-        bolt_manager::BoltManagerContract::{self, BoltManagerContractErrors},
-        request_confirmation, try_parse_contract_error,
+        // bolt_manager::BoltManagerContract::{self, BoltManagerContractErrors},
+        request_confirmation,
+        try_parse_contract_error,
     },
     contracts::{
         bolt::{
             BoltEigenLayerMiddlewareHolesky::{self, BoltEigenLayerMiddlewareHoleskyErrors},
             BoltEigenLayerMiddlewareMainnet::{self, BoltEigenLayerMiddlewareMainnetErrors},
+            BoltManager::{self, BoltManagerErrors},
             OperatorsRegistryV1::{self, OperatorsRegistryV1Errors},
             SignatureWithSaltAndExpiry,
         },
@@ -306,8 +308,7 @@ impl EigenLayerSubcommand {
                         Err(e) => parse_eigenlayer_middleware_mainnet_errors(e)?,
                     }
                 } else if chain == Chain::Holesky {
-                    let bolt_manager =
-                        BoltManagerContract::new(deployments.bolt.manager, provider.clone());
+                    let bolt_manager = BoltManager::new(deployments.bolt.manager, provider.clone());
 
                     if bolt_manager.isOperator(address).call().await?._0 {
                         info!(?address, "EigenLayer operator is registered");
@@ -330,8 +331,8 @@ impl EigenLayerSubcommand {
 
                             info!("Succesfully updated EigenLayer operator RPC");
                         }
-                        Err(e) => match try_parse_contract_error::<BoltManagerContractErrors>(e)? {
-                            BoltManagerContractErrors::OperatorNotRegistered(_) => {
+                        Err(e) => match try_parse_contract_error::<BoltManagerErrors>(e)? {
+                            BoltManagerErrors::OperatorNotRegistered(_) => {
                                 eyre::bail!("Operator not registered in bolt")
                             }
                             other => {
@@ -416,8 +417,8 @@ impl EigenLayerSubcommand {
                         Err(e) => parse_eigenlayer_middleware_mainnet_errors(e)?,
                     }
                 } else if chain == Chain::Holesky {
-                    let bolt_manager =
-                        BoltManagerContract::new(deployments.bolt.manager, provider.clone());
+                    let bolt_manager = BoltManager::new(deployments.bolt.manager, provider.clone());
+
                     if bolt_manager.isOperator(address).call().await?._0 {
                         info!(?address, "EigenLayer operator is registered");
                     } else {
@@ -434,9 +435,9 @@ impl EigenLayerSubcommand {
                         Ok(operator_data) => {
                             info!(?address, operator_data = ?operator_data._0, "Operator data");
                         }
-                        Err(e) => match try_parse_contract_error::<BoltManagerContractErrors>(e)? {
-                            BoltManagerContractErrors::KeyNotFound(_) => {
-                                warn!(?address, "Operator data not found");
+                        Err(e) => match try_parse_contract_error::<BoltManagerErrors>(e)? {
+                            BoltManagerErrors::KeyNotFound(key) => {
+                                warn!(?address, "Operator data not found: {:?}", key.key);
                             }
                             other => {
                                 unreachable!(
